Moving private functions to seperate file
